`TESSERA_LISTINGS` and `TESSERA_HOLD_STATE` rework
